requirements.txt
cltl.backend[impl,host,service]
cltl.object-recognition[impl,service]
cltl.emissor-data[impl,service,client]
cltl.chat-ui
objectref[impl,service]

flask
werkzeug
